moves the isolated files package to filestorage package.
