+++ src/main/java/nl/kb/dare/endpoints/RepositoriesEndpoint.java
@@ -1,6 +1,6 @@
 package nl.kb.dare.endpoints;
 
-import nl.kb.dare.files.FileStorage;
+import nl.kb.filestorage.FileStorage;
 import nl.kb.dare.model.oai.OaiRecordDao;
 import nl.kb.dare.model.reporting.ErrorReportDao;
 import nl.kb.dare.model.repository.Repository;
