+++ src/main/java/nl/kb/dare/oai/GetRecord.java
@@ -1,7 +1,7 @@
 package nl.kb.dare.oai;
 
-import nl.kb.dare.files.FileStorage;
-import nl.kb.dare.files.FileStorageHandle;
+import nl.kb.filestorage.FileStorage;
+import nl.kb.filestorage.FileStorageHandle;
 import nl.kb.dare.http.HttpFetcher;
 import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.manifest.ManifestFinalizer;
