+++ src/main/java/nl/kb/dare/oai/GetRecordResourceOperations.java
@@ -3,7 +3,7 @@ package nl.kb.dare.oai;
 import com.google.common.collect.Lists;
 import nl.kb.stream.ByteCountOutputStream;
 import nl.kb.stream.ChecksumOutputStream;
-import nl.kb.dare.files.FileStorageHandle;
+import nl.kb.filestorage.FileStorageHandle;
 import nl.kb.dare.http.HttpFetcher;
 import nl.kb.dare.http.HttpResponseHandler;
 import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
