+++ src/main/java/nl/kb/dare/oai/ScheduledOaiHarvester.java
@@ -3,7 +3,7 @@ package nl.kb.dare.oai;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.Lists;
 import com.google.common.util.concurrent.AbstractScheduledService;
-import nl.kb.dare.files.FileStorage;
+import nl.kb.filestorage.FileStorage;
 import nl.kb.dare.http.HttpFetcher;
 import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.model.oai.OaiRecord;
