+++ src/main/java/nl/kb/filestorage/FileStorageFactory.java
@@ -1,4 +1,4 @@
-package nl.kb.dare.files;
+package nl.kb.filestorage;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
