+++ src/main/java/nl/kb/filestorage/LocalFileStorage.java
@@ -1,4 +1,4 @@
-package nl.kb.dare.files;
+package nl.kb.filestorage;
 
 import org.apache.commons.io.FileUtils;
