+++ src/main/java/nl/kb/filestorage/LocalFileStorageHandle.java
@@ -1,4 +1,4 @@
-package nl.kb.dare.files;
+package nl.kb.filestorage;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
