+++ src/test/java/nl/kb/dare/oai/GetRecordTest.java
@@ -1,7 +1,7 @@
 package nl.kb.dare.oai;
 
 import com.google.common.collect.Lists;
-import nl.kb.dare.files.FileStorageHandle;
+import nl.kb.filestorage.FileStorageHandle;
 import nl.kb.dare.manifest.ObjectResource;
 import nl.kb.dare.model.oai.OaiRecord;
 import nl.kb.dare.model.reporting.ProgressReport;
